Melee Enemy Created
A working melee enemy is created, with animations, movement and damage dealing. The attached weapon can theoretically be attached to another entity (such as the player). It has not yet been tested to equip to another entity or with functionality as a pickup. Assets still need to be chosen for style/design purposes. Implementation of receiving damage not yet completed.